tool contracts: single DOM-free source of truth
Add client-data/tools/contracts.js as a DOM-free module merging
manifest metadata with per-tool storage contracts. Server hot paths
(message_validation, stored_svg_item_codec, canonical_items) now import
it instead of tools/index.js, so a DOM import in a tool cannot break
server startup, validation or persistence.

- Decode wire tool codes via explicit TOOL_ID_BY_CODE map membership
  instead of array-position / range checks (message_validation,
  message_tool_metadata, message_common).
- Move pencil path helpers into pencil/pencil_path.js (DOM-free).
- Add drift tests asserting manifest, runtime registry and server
  contracts agree, with literal expected wire/mutation codes and
  round trips.
- Extend boot_graph test: server contract closures exclude tool
  interaction modules.

Tool index modules now consume the shared contracts, removing ~600
lines of duplicated contract definitions.
