+++ client-data/tools/pencil/index.js
@@ -27,7 +27,14 @@
 import { logFrontendEvent } from "../../js/frontend_logging.js";
 import { LIMITS } from "../../js/message_common.js";
 import { MutationType } from "../../js/mutation_type.js";
+import { PencilContract } from "../contracts.js";
 import { TOOL_CODE_BY_ID } from "../tool-order.js";
+import {
+  appendPersistedPencilPath,
+  renderPencilPath,
+  scanPathSummary,
+  serializeStoredPencilPath,
+} from "./pencil_path.js";
 import { wboPencilPoint } from "./wbo_pencil_point.js";
 
 /** @import { ToolBootContext, ToolRuntimeModules } from "../../../types/app-runtime" */
@@ -62,220 +69,6 @@ function isPencilMessage(data) {
   );
 }
 
-/**
- * @param {number} value
- * @returns {number}
- */
-function roundPathValue(value) {
-  return Math.round(value);
-}
-
-/**
- * Parse a canonical persisted pencil path emitted by renderPencilPath:
- * `M <x> <y> l <dx> <dy> ...`
- *
- * @param {string} d
- * @param {number} index
- * @returns {{value: number, index: number} | null}
- */
-function readCanonicalPathInteger(d, index) {
-  const length = d.length;
-  let next = index;
-  let sign = 1;
-  if (next < length && d.charCodeAt(next) === 45) {
-    sign = -1;
-    next += 1;
-  }
-  if (next >= length) return null;
-  let code = d.charCodeAt(next) - 48;
-  if (code < 0 || code > 9) return null;
-  let value = 0;
-  while (next < length && code >= 0 && code <= 9) {
-    value = value * 10 + code;
-    next += 1;
-    if (next >= length) break;
-    code = d.charCodeAt(next) - 48;
-  }
-  return { value: sign * value, index: next };
-}
-
-/**
- * @param {string | undefined} d
- * @returns {{
- *   childCount: number,
- *   localBounds: {minX: number, minY: number, maxX: number, maxY: number} | null,
- *   lastPoint: {x: number, y: number} | null,
- * }}
- */
-const EMPTY_PERSISTED_PENCIL_SCAN = {
-  childCount: 0,
-  localBounds: null,
-  lastPoint: null,
-};
-
-/**
- * @param {string | undefined} d
- * @returns {{
- *   childCount: number,
- *   localBounds: {minX: number, minY: number, maxX: number, maxY: number} | null,
- *   lastPoint: {x: number, y: number} | null,
- * }}
- */
-function scanPersistedPencilPath(d) {
-  if (typeof d !== "string" || d === "") return EMPTY_PERSISTED_PENCIL_SCAN;
-  const length = d.length;
-  if (length < 5 || d.charCodeAt(0) !== 77 || d.charCodeAt(1) !== 32)
-    return EMPTY_PERSISTED_PENCIL_SCAN;
-
-  let index = 2;
-  const firstX = readCanonicalPathInteger(d, index);
-  if (!firstX || firstX.index >= length || d.charCodeAt(firstX.index) !== 32)
-    return EMPTY_PERSISTED_PENCIL_SCAN;
-  index = firstX.index + 1;
-  const firstY = readCanonicalPathInteger(d, index);
-  if (!firstY) return EMPTY_PERSISTED_PENCIL_SCAN;
-  index = firstY.index;
-
-  let currentX = firstX.value;
-  let currentY = firstY.value;
-  let minX = currentX;
-  let minY = currentY;
-  let maxX = currentX;
-  let maxY = currentY;
-  let childCount = 1;
-  let previousDistinctX = currentX;
-  let previousDistinctY = currentY;
-
-  while (index < length) {
-    if (
-      index + 2 >= length ||
-      d.charCodeAt(index) !== 32 ||
-      d.charCodeAt(index + 1) !== 108 ||
-      d.charCodeAt(index + 2) !== 32
-    ) {
-      return EMPTY_PERSISTED_PENCIL_SCAN;
-    }
-    index += 3;
-    const deltaX = readCanonicalPathInteger(d, index);
-    if (!deltaX || deltaX.index >= length || d.charCodeAt(deltaX.index) !== 32)
-      return EMPTY_PERSISTED_PENCIL_SCAN;
-    index = deltaX.index + 1;
-    const deltaY = readCanonicalPathInteger(d, index);
-    if (!deltaY) return EMPTY_PERSISTED_PENCIL_SCAN;
-    index = deltaY.index;
-    currentX += deltaX.value;
-    currentY += deltaY.value;
-    if (previousDistinctX === currentX && previousDistinctY === currentY) {
-      continue;
-    }
-    previousDistinctX = currentX;
-    previousDistinctY = currentY;
-    childCount += 1;
-    if (currentX < minX) minX = currentX;
-    else if (currentX > maxX) maxX = currentX;
-    if (currentY < minY) minY = currentY;
-    else if (currentY > maxY) maxY = currentY;
-  }
-
-  return {
-    childCount,
-    localBounds: { minX, minY, maxX, maxY },
-    lastPoint: { x: currentX, y: currentY },
-  };
-}
-
-/**
- * @param {string | undefined} d
- * @returns {{childCount: number, localBounds: {minX: number, minY: number, maxX: number, maxY: number} | null}}
- */
-function scanPathSummary(d) {
-  const scanned = scanPersistedPencilPath(d);
-  return {
-    childCount: scanned.childCount,
-    localBounds: scanned.localBounds,
-  };
-}
-
-/**
- * @param {string | undefined} d
- * @param {{x: number, y: number}[]} points
- * @returns {string}
- */
-function appendPersistedPencilPath(d, points) {
-  if (!Array.isArray(points) || points.length === 0) {
-    return typeof d === "string" ? d : "";
-  }
-  const scanned = scanPersistedPencilPath(d);
-  if (!scanned.lastPoint) return "";
-  let lastX = scanned.lastPoint.x;
-  let lastY = scanned.lastPoint.y;
-  let pathData = d || "";
-  for (let index = 0; index < points.length; index += 1) {
-    const point = points[index];
-    if (!point || !Number.isFinite(point.x) || !Number.isFinite(point.y)) {
-      continue;
-    }
-    const x = roundPathValue(point.x);
-    const y = roundPathValue(point.y);
-    pathData += ` l ${x - lastX} ${y - lastY}`;
-    lastX = x;
-    lastY = y;
-  }
-  return pathData;
-}
-
-/**
- * @param {{x: number, y: number}[]} points
- * @returns {string}
- */
-function renderPencilPath(points) {
-  if (!Array.isArray(points) || points.length === 0) return "";
-  const firstPoint = points[0];
-  if (
-    !firstPoint ||
-    !Number.isFinite(firstPoint.x) ||
-    !Number.isFinite(firstPoint.y)
-  ) {
-    return "";
-  }
-  let lastX = roundPathValue(firstPoint.x);
-  let lastY = roundPathValue(firstPoint.y);
-  let pathData = `M ${lastX} ${lastY}`;
-  if (points.length === 1) return `${pathData} l 0 0`;
-  for (let index = 1; index < points.length; index += 1) {
-    const point = points[index];
-    if (!point || !Number.isFinite(point.x) || !Number.isFinite(point.y)) {
-      continue;
-    }
-    const x = roundPathValue(point.x);
-    const y = roundPathValue(point.y);
-    pathData += ` l ${x - lastX} ${y - lastY}`;
-    lastX = x;
-    lastY = y;
-  }
-  return pathData;
-}
-
-/**
- * @param {StoredPencilPathItem} item
- * @param {string} pathData
- * @param {StoredPencilPathSerializeHelpers} helpers
- * @returns {string}
- */
-function serializeStoredPencilPath(item, pathData, helpers) {
-  if (!pathData) return "";
-  const transform = helpers.renderTransformAttribute(item.transform);
-  const id = typeof item.id === "string" ? helpers.escapeHtml(item.id) : "";
-  const color = helpers.escapeHtml(item.color || "#000000");
-  const size = helpers.numberOrZero(item.size) | 0;
-  const opacity =
-    typeof item.opacity === "number" ? ` opacity="${item.opacity}"` : "";
-  return (
-    `<path id="${id}" d="${helpers.escapeHtml(pathData)}"` +
-    ` stroke="${color}" stroke-width="${size}" fill="none" stroke-linecap="round" stroke-linejoin="round"${opacity}${transform}></path>`
-  );
-}
-
 export {
   appendPersistedPencilPath,
   renderPencilPath,
@@ -287,57 +80,7 @@ export const toolId = "pencil";
 const toolCode = TOOL_CODE_BY_ID[toolId];
 export const drawsOnBoard = true;
 
-/** @type {import("../shape_contract.js").ToolContract} */
-const contract = {
-  toolId,
-  toolCode,
-  payloadKind: "children",
-  storedTagName: "path",
-  liveMessageFields: /** @type {const} */ ({
-    [MutationType.CREATE]: {
-      id: "id",
-      color: "color",
-      size: "size",
-      opacity: "opacity?",
-    },
-    [MutationType.APPEND]: {
-      parent: "id",
-      x: "coord",
-      y: "coord",
-    },
-  }),
-  summarizeStoredSvgItem(entry, paintOrder, helpers) {
-    const size = helpers.parseNumber(
-      helpers.readStoredSvgAttribute(entry, "stroke-width"),
-    );
-    const scanned = scanPathSummary(helpers.readStoredSvgAttribute(entry, "d"));
-    if (size === undefined || scanned.childCount === 0) return null;
-    return {
-      id: helpers.id,
-      tool: contract.toolId,
-      data: helpers.decorateStoredItemData(
-        {
-          color: helpers.readStoredSvgAttribute(entry, "stroke") || "#000000",
-          size,
-        },
-        helpers.opacity,
-        helpers.transform,
-      ),
-      childCount: scanned.childCount,
-      paintOrder,
-      localBounds: scanned.localBounds,
-    };
-  },
-  serializeStoredSvgItem(item, helpers) {
-    const points = Array.isArray(item._children) ? item._children : [];
-    const pathData = renderPencilPath(points);
-    return serializeStoredPencilPath(item, pathData, helpers);
-  },
-  renderBoardSvg(pencil, helpers) {
-    const pathstring = renderPencilPath(pencil._children || []);
-    return helpers.renderPath(pencil, pathstring);
-  },
-};
+const contract = PencilContract;
 
 export { contract };
 export const shortcut = "p";
